backend/requirements.txt
# Core dependencies
streamlit==1.28.0

# FastAPI for REST API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - MySQL support (SQLite is built-in to Python)
mysql-connector-python==8.2.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# Environment & Configuration
python-dotenv==1.0.0

# CORS support
fastapi-cors==0.0.6

# Date/Time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Logging
colorlog==6.8.0

# Async support
aiofiles==23.2.1
httpx==0.25.2

# Data validation
pydantic==2.5.2
email-validator==2.1.0

# Development tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

